+++ examples/logic/config.rb
@@ -15,3 +15,4 @@ javascripts_dir = "js"
 
 # To disable debugging comments that display the original location of your selectors. Uncomment:
 line_comments = false
+